Add files via upload
This is SQL chunk code for the dataset obtained from the US website, Mother Jones, listing some of the most infamous mass shootings in US history from the 80's to present day. The original dataset was downloaded from the aforementioned website and manipulated to fit the analysis executed in the SQL codes. See reference links below for both my Google Sheets data manipulation and the main data source.  An MS Excel would be added to this  project for exploration by non-Sheets users.

https://docs.google.com/spreadsheets/d/14I0BgkXh1QdRUbCzaBqhvOzCvQQgaffK67mE38GzoIE/edit?usp=sharing

https://www.motherjones.com/politics/2012/12/mass-shootings-mother-jones-full-data/